problem area classes are combined in single module.
