+++ tests/integration/test_repair_tools.py
@@ -7,7 +7,8 @@ from ansys.geometry.core.tools.repair_tools import RepairTools
 def test_find_split_edges():
     modeler = Modeler(host="localhost", port=50051)
     modeler.open_file("./tests/integration/files/SplitEdgeDesignTest.scdoc")
-    problem_areas = modeler.tools.prepare.FindSplitEdges(["0:39"], 25, 150)
+    rep = RepairTools()
+    problem_areas = rep.find_split_edges(["0:39"], 25, 150)
     assert len(problem_areas) == 3
 
 
@@ -16,7 +17,7 @@ def test_find_extra_edges():
     modeler = Modeler(host="localhost", port=50051)
     modeler.open_file("./tests/integration/files/ExtraEdgesDesignBefore.scdoc")
     rep = RepairTools()
-    problem_areas = rep.FindExtraEdges(["0:22"])
+    problem_areas = rep.find_extra_edges(["0:22"])
     assert len(problem_areas) == 1
 
 
@@ -25,7 +26,7 @@ def test_find_inexact_edges():
     modeler = Modeler(host="localhost", port=50051)
     modeler.open_file("./tests/integration/files/InExactEdgesBefore.scdoc")
     rep = RepairTools()
-    problem_areas = rep.FindInexactEdges(["0:38"])
+    problem_areas = rep.find_inexact_edges(["0:38"])
     assert len(problem_areas) == 6
 
 
@@ -34,7 +35,7 @@ def test_find_missing_faces():
     modeler = Modeler(host="localhost", port=50051)
     modeler.open_file("./tests/integration/files/MissingFacesDesignBefore.scdoc")
     rep = RepairTools()
-    problem_areas = rep.FindMissingFaces(["1:40"])
+    problem_areas = rep.find_missing_faces(["1:40"])
     assert len(problem_areas) == 1
 
 
@@ -43,7 +44,7 @@ def test_find_duplicate_faces():
     modeler = Modeler(host="localhost", port=50051)
     modeler.open_file("./tests/integration/files/DuplicateFacesDesignBefore.scdoc")
     rep = RepairTools()
-    problem_areas = rep.FindDuplicateFaces(["0:22", "0:85"])
+    problem_areas = rep.find_duplicate_faces(["0:22", "0:85"])
     assert len(problem_areas) == 1
 
 
@@ -52,7 +53,7 @@ def test_find_small_faces():
     modeler = Modeler(host="localhost", port=50051)
     modeler.open_file("./tests/integration/files/SmallFacesBefore.scdoc")
     rep = RepairTools()
-    problem_areas = rep.FindSmallFaces(["0:38"])
+    problem_areas = rep.find_small_faces(["0:38"])
     assert len(problem_areas) == 4
 
 
@@ -65,5 +66,5 @@ def test_find_stitch_faces():
     for body in design.bodies:
         faceIds.append(body.id)
     rep = RepairTools()
-    problem_areas = rep.FindStitchFaces(faceIds)
+    problem_areas = rep.find_stitch_faces(faceIds)
     assert len(problem_areas) == 1
